+++ R/makeQset.R
@@ -135,7 +135,7 @@ makeQset <- function(sampleTable,
 
     #this is included in the PairedAndR1s method more efficiently, don't need to call it there.
     #note this also adds the input relH if input files provided.
-    qseaSet <- addMedipsEnrichmentFactors(qseaSet, nCores = BiocParallel::bpworkers())
+    qseaSet <- addMedipsEnrichmentFactors(qseaSet, nCores = BiocParallel::bpworkers(), nonEnrich = FALSE)
 
   } else if (coverageMethod == "PairedAndR1s") {